# GreekLearningApp-Frontend/swagger.json
{
...
  "openapi": "3.0.0",
  "info": {
    "title": "Greek Parser API",
    "description": "API for Greek language learning and parsing",
    "version": "v1"
  },
  "servers": [
    {
      "url": "http://localhost:5001"
    }
  ],
  "paths": {
...
    "/api/words": {
      "get": {
        "tags": [
          "Vocabulary"
        ],
        "operationId": "Vocabulary_GetAll",
        "parameters": [
          {
            "name": "occurances",
            "in": "query",
            "schema": {
              "type": "integer",
              "format": "int32",
              "nullable": true
            },
            "x-position": 1
          },
          {
            "name": "comparison",
            "in": "query",
            "schema": {
              "type": "string",
              "nullable": true
            },
            "x-position": 2
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/SimpleWordDto"
                  }
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "Vocabulary"
        ],
        "operationId": "Vocabulary_Create",
        "requestBody": {
          "x-name": "createDto",
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateVocabularyDto"
              }
            }
          },
          "required": true,
          "x-position": 1
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/VocabularyDto"
                }
              }
            }
          }
        }
      }
    },
    "/api/words/{id}": {
      "get": {
        "tags": [
          "Vocabulary"
        ],
        "operationId": "Vocabulary_GetById",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            },
            "x-position": 1
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SimpleWordDto"
                }
              }
            }
          }
        }
      },
      "put": {
        "tags": [
          "Vocabulary"
        ],
        "operationId": "Vocabulary_Update",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            },
            "x-position": 1
          }
        ],
        "requestBody": {
          "x-name": "updateDto",
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateVocabularyDto"
              }
            }
          },
          "required": true,
          "x-position": 2
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/VocabularyDto"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "Vocabulary"
        ],
        "operationId": "Vocabulary_Delete",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            },
            "x-position": 1
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/octet-stream": {
                "schema": {
                  "type": "string",
                  "format": "binary"
                }
              }
            }
          }
        }
      }
    },
    "/api/words/update": {
      "post": {
        "tags": [
          "Vocabulary"
        ],
        "operationId": "Vocabulary_UpdateWord",
        "requestBody": {
          "x-name": "word",
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SimpleWordDto"
              }
            }
          },
          "required": true,
          "x-position": 1
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/octet-stream": {
                "schema": {
                  "type": "string",
                  "format": "binary"
                }
              }
            }
          }
        }
      }
    },
    "/api/words/search": {
      "get": {
        "tags": [
          "Vocabulary"
        ],
        "operationId": "Vocabulary_Search",
        "parameters": [
          {
            "name": "SearchTerm",
            "in": "query",
            "schema": {
              "type": "string",
              "nullable": true
            },
            "x-position": 1
          },
          {
            "name": "PartOfSpeech",
            "in": "query",
            "schema": {
              "type": "string",
              "nullable": true
            },
            "x-position": 2
          },
          {
            "name": "Page",
            "in": "query",
            "schema": {
              "type": "integer",
              "format": "int32"
            },
            "x-position": 3
          },
          {
            "name": "PageSize",
            "in": "query",
            "schema": {
              "type": "integer",
              "format": "int32"
            },
            "x-position": 4
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PaginatedResultDtoOfVocabularyDto"
                }
              }
            }
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
...
      "SimpleWordDto": {
        "type": "object",
        "additionalProperties": false,
        "properties": {
          "rootGUID": {
            "type": "string",
            "format": "guid"
          },
          "rootId": {
            "type": "integer",
            "format": "int32"
          },
          "content": {
            "type": "string"
          },
          "occurances": {
            "type": "integer",
            "format": "int32"
          },
          "gloss": {
            "type": "string",
            "nullable": true
          }
        }
      },
...
      "PaginatedResultDtoOfVocabularyDto": {
        "type": "object",
        "additionalProperties": false,
        "properties": {
          "items": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/VocabularyDto"
            }
          },
          "totalCount": {
            "type": "integer",
            "format": "int32"
          },
          "page": {
            "type": "integer",
            "format": "int32"
          },
          "pageSize": {
            "type": "integer",
            "format": "int32"
          },
          "totalPages": {
            "type": "integer",
            "format": "int32"
          }
        }
      },
      "VocabularyDto": {
        "type": "object",
        "additionalProperties": false,
        "properties": {
          "id": {
            "type": "integer",
            "format": "int32"
          },
          "root": {
            "type": "string"
          },
          "transliteration": {
            "type": "string",
            "nullable": true
          },
          "gloss": {
            "type": "string"
          },
          "partOfSpeech": {
            "type": "string",
            "nullable": true
          },
          "frequencyRank": {
            "type": "integer",
            "format": "int32",
            "nullable": true
          }
        }
      },
      "CreateVocabularyDto": {
        "type": "object",
        "additionalProperties": false,
        "properties": {
          "root": {
            "type": "string"
          },
          "transliteration": {
            "type": "string",
            "nullable": true
          },
          "gloss": {
            "type": "string"
          },
          "partOfSpeech": {
            "type": "string",
            "nullable": true
          },
          "frequencyRank": {
            "type": "integer",
            "format": "int32",
            "nullable": true
          }
        }
      },
      "UpdateVocabularyDto": {
        "type": "object",
        "additionalProperties": false,
        "properties": {
          "root": {
            "type": "string"
          },
          "transliteration": {
            "type": "string",
            "nullable": true
          },
          "gloss": {
            "type": "string"
          },
          "partOfSpeech": {
            "type": "string",
            "nullable": true
          },
          "frequencyRank": {
            "type": "integer",
            "format": "int32",
            "nullable": true
          }
        }
      }
    },
    "securitySchemes": {
      "Bearer": {
        "type": "apiKey",
        "description": "Type into the textbox: Bearer {your JWT token}.",
        "name": "Authorization",
        "in": "header"
      }
    }
  }
}